Moved FSM to new folder
